backend/requirements.txt
bcrypt==4.0.1
email-validator==1.3.0
fastapi==0.88.0
greenlet==2.0.1
passlib==1.7.4
Pillow==9.3.0
pydantic==1.10.2
PyJWT==2.6.0
PyMySQL==1.0.2
python-dotenv==0.21.0
python-multipart==0.0.5
SQLAlchemy==1.4.45
uvicorn==0.20.0
psycopg2